Rename the main package in test directory
This commit fixes a test initialization error introduced in the commit
179e7c and discovered just now. I missed the error before because I
hadn't written any tests at the moment, so I didn't feel like I had to
run the one test that was generated by Spring Tool Suite while
creating the project.

In the commit introducing the error, I renamed the main package of the
application from:

com.github.piotr.rusin.yule

to

com.github.piotr_rusin.yule,

but I forgot to do the same for its corresponding package in src/test
directory of the project.

As a result, no application configuration could be found, both for the
generated tests and for those introduced by me.