Remove PRIVATE status
The PRIVATE article status was intended for pages and blog posts that
were to be accessible in the same way as any public page or blog post
(that is: through the index page, links and URLs using page or blog post
URL format), but only for users logged in as administrators (that is: so
far for anyone with login credentials).

With the support fo multiple administrative users removed, keeping this
feature and plans for it doesn't make much sense.