Replace status property of Article class
This commit:
- removes the property
- removes ArticleStatus enum
- adds published property to Article
- renames methods of ArticleRepository interface and their integration
  tests to make them refer published field instead of the PUBLIC
  status

The main reasons behind these changes is that, after removing the
PRIVATE status (combining information that the article is published but
visible only to authenticated users), I realised expressing status as
one of values of an enum is no longer necessary and expressing both the
remaining statuses (PUBLIC and DRAFT) and some planned ones would be
simpler with separate fields or combinations of fields.

Another advantage is that now changing a status will require only
changing necessary information pertaining to it (like value of published
and publicationDate fields in case of a future publication status, or
value of a "removed" field in case of a removed article status), while
the rest of the information pertaining to the previous status of the
article will be preserved.