Implement constructors for Article class
This commit adds two constructors: a protected, no-argument one and a
public one requiring a title and content of the new article.

The second constructor, combined with all the methods of the Article
class, will ensure an article object created by using the new operator
will have valid state.

The first one is required by JPA and is to be used by Hibernate.