Simplify Article.setStatus(ArticleStatus)
The implementation of the method is being simplified - it is no longer
concerned with leaving the instance of Article it is called on in a
valid state. Validation of instances of the class will be handled
separately by a java bean validation provider.

Adding validation rules in form of annotations will not only simplify
the code, but also allow me to use instances of the class as form
objects, so that they could be populated with valid or invalid data and
validated before being passed to controller methods.