Implement caching of previous state of an Article
This commit adds a private "cache" field to the Article class and two
methods: Article.cache() and Article.getCache(). The cache method is
responsible for storing current state of the article in the "cache"
field as a separate instance of the Article class. It is annotated as a
post-load event handler.

This way, the initial, post-load state of an article will be easily
available.