Make sure introduction isn't the whole content
This commit removes behaviour of Article.getIntroduction() in which, in
case no custom introduction is assigned to an article, the first element
of the result of splitting the article content by the "<!--more-->" tag
is always treated as an introduction, even if there is no tag and the
returned element happens to be the whole article.

Now in such cases, a null value will be returned and the code expecting
the introduction value will be able to react appropriately.