Replace a handler for index page request
The handler method is being replaced with one that handles a request for
a page of a list of published blog posts: the first one when request
path matches "/", and the requested one when it contains a page number
as a path variable.

The request is still being handled by returning the same index.html
template as before. The template will be updated soon.