Fix auto-publication rescheduling mechanism
Saving a new article as scheduled for publication didn't trigger
re-scheduling of auto-publication task because a call to
AutoPublicationScheduler.scheduleNew() was missing.

Updating an article never triggered the re-scheduling operation because
instead of using the copy of article from before the update to determine
if re-scheduling is necessary, the code used a copy made after the
update.