Fix first page link in the pagination
The first page link always pointed to the first page of index page. This
commit fixes that.

In addition, the pagination fragment layout is being reorganized. Now it
requires only the first, paginated-content-specific component of the
path to be passed as its parameter. The page specific component - a
template string with page number as its variable - can still be passed,
but now it's an optional parameter.