Add currently edited Article to SessionAttributes
This change makes it possible to use Article form object merged with
session-stored instance of Article to update an existing article object,
instead of having to query for the existing object and update it with
data from the form object.

It also fixes an error occuring when early return due to detection of
validation errors results in treating the form object (unmerged with
then non-existent session-stored value) as a new instance of Article,
which in turn results in duplicate title errors when attempting to save
it.

I didn't store the article in session before because I wanted to use
REST-inspired URLs that would uniquely identify Articles, and since I
could identify the article from a URL, storing it in session seemed
redundant.

Previously, the query for existing article was not only necessary to get
an Article instance identified by URL, but it was also necessary due to
more complex auto-publication rescheduling mechanism that avoided
rescheduling when it was not necessary - for this, I needed to compare
the article in the database with the form object to detect changes in
article status or in scheduled publication date.